+++ HopStepJumpJump/Util/Ranking.cpp
@@ -154,7 +154,7 @@ std::string Ranking::HttpGet(const char* domain, const char* uri)
 #endif // _DEBUG
 
 		// データ送信(http命令を送る)
-		NetWorkSend(m_netH, HttpCmd, strlen(HttpCmd));
+		NetWorkSend(m_netH, HttpCmd, static_cast<int>(strlen(HttpCmd)));
 
 		// データがくるのを待つ
 		while (!ProcessMessage())
